refactoring and add method addPostLike,getCommentLikes

+++ server/src/controllers/likeController.js
@@ -71,9 +71,16 @@ export async function addCommentLike(req, res) {
         .json({ message: `Comment with ID ${commentId} was not found` });
     }
 
+    if (comment.post.toString() !== postId) {
+      return res.status(400).json({
+        message: `The comment with ID ${commentId} does not belong to the post with ID ${postId}`,
+      });
+    }
+
     const existingLike = await CommentLike.findOne({
       user: userId,
       comment: commentId,
+      post: postId,
     });
     if (existingLike) {
       await CommentLike.findByIdAndDelete(existingLike._id);
@@ -88,11 +95,82 @@ export async function addCommentLike(req, res) {
     const newLike = new CommentLike({
       user: userId,
       comment: commentId,
+      post: postId,
     });
     await newLike.save();
     user.commentLikes.push(newLike._id);
     await user.save();
     return res.status(200).json({ message: "Like added to comment" });
+  } catch (error) {
+    res.status(500).json({ message: "Server internal error", error: error });
+  }
+}
+
+export async function getPostlikes(req, res) {
+  try {
+    const { postId } = req.params;
+    const post = await Post.findById(postId);
+    if (!post) {
+      return res
+        .status(400)
+        .json({ message: `post with id ${postId} was not found` });
+    }
+    const postslikeCount = await PostLike.countDocuments({ post: postId });
+    if (postslikeCount === 0) {
+      return res.status(400).json({
+        message: `No likes found for the post with id ${postId}`,
+      });
+    }
+    res.status(200).json({
+      message: `was found ${postslikeCount} like for Post with id ${postId}`,
+      postslikeCount,
+    });
+  } catch (error) {
+    res.status(500).json({ message: "Server internal error", error: error });
+  }
+}
+
+export async function getCommentLikes(req, res) {
+  try {
+    const { postId, commentId } = req.params;
+    const userId = req.user.id;
+    const user = await User.findById(userId);
+
+    if (!user) {
+      return res.status(404).json({ message: "user was not found" });
+    }
+
+    const post = await Post.findById(postId);
+    if (!post) {
+      return res
+        .status(400)
+        .json({ message: `post with id ${postId} was not found` });
+    }
+    const commentsLikeCount = await CommentLike.countDocuments({
+      comment: commentId,
+    });
+    const comment = await Comment.findById(commentId);
+    if (!comment) {
+      return res
+        .status(400)
+        .json({ message: `comment with id ${commentId} was not found` });
+    }
+
+    if (comment.post.toString() !== postId) {
+      return res.status(400).json({
+        message: `The comment with ID ${commentId} does not belong to the post with ID ${postId}`,
+      });
+    }
+
+    if (commentsLikeCount === 0) {
+      return res.status(400).json({
+        message: `No likes found for the comment with id ${commentId}`,
+      });
+    }
+    res.status(200).json({
+      message: `was found ${commentsLikeCount} like for comment with id ${commentId}`,
+      commentsLikeCount,
+    });
   } catch (error) {
     res.status(500).json({ message: "Server internal error", error: error });
   }
